+++ examples/Custom_Surrogates/custom_architecture_torch.py
@@ -21,8 +21,8 @@ from baybe.parameters import (
     SubstanceParameter,
 )
 from baybe.recommenders import (
+    BotorchRecommender,
     FPSRecommender,
-    SequentialGreedyRecommender,
     TwoPhaseMetaRecommender,
 )
 from baybe.searchspace import SearchSpace
@@ -190,9 +190,7 @@ campaign = Campaign(
     searchspace=SearchSpace.from_product(parameters=parameters, constraints=None),
     objective=SingleTargetObjective(target=NumericalTarget(name="Yield", mode="MAX")),
     recommender=TwoPhaseMetaRecommender(
-        recommender=SequentialGreedyRecommender(
-            surrogate_model=NeuralNetDropoutSurrogate()
-        ),
+        recommender=BotorchRecommender(surrogate_model=NeuralNetDropoutSurrogate()),
         initial_recommender=FPSRecommender(),
     ),
 )