+++ tests/test_acquisition.py
@@ -2,7 +2,7 @@
 
 from hypothesis import given
 
-from baybe.recommenders import SequentialGreedyRecommender
+from baybe.recommenders import BotorchRecommender
 
 from .hypothesis_strategies.acquisition import acquisition_functions
 
@@ -10,4 +10,4 @@ from .hypothesis_strategies.acquisition import acquisition_functions
 @given(acquisition_functions)
 def test_acqfs(acqf):
     """Test all acquisition functions with sequential greedy recommender."""
-    SequentialGreedyRecommender(acquisition_function=acqf)
+    BotorchRecommender(acquisition_function=acqf)